requirements.txt
numpy~=1.21.5
scipy~=1.7.0
Pillow~=8.2.0
torch~=2.1.0
torchvision~=0.16.0
matplotlib~=3.7.2
tqdm~=4.61.1
opencv-python~=4.5.3.56
scikit-learn~=0.24.2
colorama~=0.4.4
pykeops~=2.1
pyyaml~=6.0

pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu121
